Rename QueryData to ManageData + Use LoadingScreen instead if h1 text
