+++ src/App.jsx
@@ -6,8 +6,7 @@ import Logout from './components/Logout';
 import Layout from './components/Layout/Layout';
 import Dashboard from './pages/Dashboard/Dashboard';
 import Profile from './pages/Profile/Profile';
-import Data from './pages/Data/Data';
-import QueryData from './pages/QueryData/QueryData';
+import ManageData from './pages/ManageData/ManageData';
 import AddData from './pages/AddData/AddData';
 import ProtectedRoute from './utils/ProtectedRoute';
 import RegisterPage from './pages/Register/RegisterPage';
@@ -58,7 +57,8 @@ function App() {
               path="/manage-attributes"
               element={
                 <ProtectedRoute
-                  Component={Data}
+                  // TODO: Replace ManageData with ManageAttributes
+                  Component={ManageData}
                   redirectPath="/login"
                   roles={[ADMIN_ROLE, VIEWER_ROLE, EDITOR_ROLE]}
                 />
@@ -70,19 +70,7 @@ function App() {
               path="/manage-data"
               element={
                 <ProtectedRoute
-                  Component={Data}
-                  redirectPath="/login"
-                  roles={[ADMIN_ROLE, VIEWER_ROLE, EDITOR_ROLE]}
-                />
-              }
-            />
-
-            <Route
-              exact
-              path="/query-data"
-              element={
-                <ProtectedRoute
-                  Component={QueryData}
+                  Component={ManageData}
                   redirectPath="/login"
                   roles={[ADMIN_ROLE, VIEWER_ROLE, EDITOR_ROLE]}
                 />
