+++ src/components/Login/Login.jsx
@@ -2,6 +2,8 @@ import React, { useState, useEffect } from 'react';
 import { Form, Input, Button, Typography } from 'antd';
 import { PropTypes, instanceOf } from 'prop-types';
 import { Navigate } from 'react-router-dom';
+
+import LoadingScreen from '../../common/LoadingScreen/LoadingScreen';
 import { GSPBackend } from '../../utils/utils';
 import { withCookies, cookieKeys, Cookies, clearCookies } from '../../utils/cookie_utils';
 import { logInWithEmailAndPassword, useNavigate, refreshToken } from '../../utils/auth_utils';
@@ -53,7 +55,7 @@ const Login = ({ roles, cookies }) => {
   };
 
   if (isLoading) {
-    return <h1>LOADING...</h1>;
+    return <LoadingScreen />;
   }
 
   if (isAuthenticated) {
