+++ src/pages/ManageData/ManageData.jsx
@@ -4,13 +4,13 @@ import { Utils as QbUtils, Query, Builder, AntdConfig } from '@react-awesome-que
 import { Alert, Button } from 'antd';
 import '@react-awesome-query-builder/antd/css/styles.css';
 
-import LoadingScreen from '../../common/LoadingScreen';
+import LoadingScreen from '../../common/LoadingScreen/LoadingScreen';
 import SelectAttributesModal from './SelectAttributesModal/SelectAttributesModal';
 import QueryResults from './QueryResults/QueryResults';
-import { tableToWidget } from './QueryDataUtils';
+import { tableToWidget } from './ManageDataUtils';
 import { GSPBackend } from '../../utils/utils';
 
-import styles from './QueryData.module.css';
+import styles from './ManageData.module.css';
 
 class DefaultDict {
   constructor(DefaultInit) {
@@ -29,7 +29,7 @@ class DefaultDict {
 
 const queryValue = { id: QbUtils.uuid(), type: 'group' };
 
-const QueryData = () => {
+const ManageData = () => {
   const [isLoading, setIsLoading] = useState(true);
   const [config, setConfig] = useState({ ...AntdConfig, fields: {} });
   const [tableState, setTableState] = useState({
@@ -41,7 +41,7 @@ const QueryData = () => {
     config,
     results: [],
   });
-  const [errorState, setErrorState] = useState("");
+  const [errorState, setErrorState] = useState('');
   const [checkedLists, setCheckedLists] = useState(new DefaultDict(Array));
 
   const [isSelectAttributesModalOpen, setIsSelectedAttributesModalOpen] = useState(false);
@@ -53,11 +53,11 @@ const QueryData = () => {
 
   const onAdvancedSearch = async () => {
     // only query if checkedLists is nonempty
-    if(Object.values(checkedLists).every((arr) => (arr.length === 0))){
-      setErrorState("Please select at least one attribute to query");
+    if (Object.values(checkedLists).every(arr => arr.length === 0)) {
+      setErrorState('Please select at least one attribute to query');
       return;
     }
-    setErrorState("");
+    setErrorState('');
 
     // get rid of prev. results (find more elegant way to do this?)
     setQueryState(prevState => ({ ...prevState, results: [] }));
@@ -72,7 +72,7 @@ const QueryData = () => {
       });
 
       setQueryState(prevState => ({ ...prevState, results: results.data }));
-    } catch(err){
+    } catch (err) {
       setErrorState(`${err.name}: ${err.message}`);
     }
   };
@@ -148,11 +148,11 @@ const QueryData = () => {
       </div>
       {
         // temporary error banner
-        (errorState !== "") &&  <Alert message={errorState} type="error" showIcon />
+        errorState !== '' && <Alert message={errorState} type="error" showIcon />
       }
       <QueryResults data={queryState.results} />
     </div>
   );
 };
 
-export default QueryData;
+export default ManageData;
