+++ src/pages/ManageData/ManageData.module.css
@@ -4,13 +4,13 @@
 }
 
 /* Query Results */
-.query-results-header{
+.query-results-header {
   margin: 2.5rem 0 2rem 0;
   display: flex;
   flex-direction: row;
   justify-content: space-between;
   align-items: flex-start;
 }
-.query-results-header h2{
+.query-results-header h2 {
   margin: 0;
 }
