+++ src/pages/ManageData/QueryResults/QueryResults.jsx
@@ -2,15 +2,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { Table, Space, Button } from 'antd';
-import { TABLE_PRIMARY_KEYS } from '../QueryDataUtils';
+import { TABLE_PRIMARY_KEYS } from '../ManageDataUtils';
 
-import styles from "./QueryResults.module.css";
+import styles from './QueryResults.module.css';
 
 const computeColumns = data => {
   if (data.length === 0 || data[0].length <= TABLE_PRIMARY_KEYS.length) {
     return [{}];
   }
-  const cols = Object.keys(data[0]).filter((value)=>(!TABLE_PRIMARY_KEYS.includes(value)))
+  const cols = Object.keys(data[0]).filter(value => !TABLE_PRIMARY_KEYS.includes(value));
 
   const actionColumn = {
     title: '',
@@ -36,10 +36,12 @@ const QueryResults = ({ data }) => {
   return (
     <div className={styles['query-results-container']}>
       <Table
-        rowKey={record => {return TABLE_PRIMARY_KEYS.map((value) => record[value]);} }
+        rowKey={record => {
+          return TABLE_PRIMARY_KEYS.map(value => record[value]);
+        }}
         dataSource={data}
         columns={computeColumns(data)}
-        scroll={{x: true}}
+        scroll={{ x: true }}
         size="middle"
       />
     </div>
