+++ src/pages/ManageData/QueryResults/QueryResults.module.css
@@ -1,0 +1,3 @@
+.query-results-container {
+  max-width: 95%;
+}
