+++ src/pages/ManageData/SelectAttributesModal/AttributeGroup/AttributeGroup.css
