+++ src/pages/ManageData/SelectAttributesModal/AttributeGroup/AttributeGroup.jsx
