+++ src/pages/ManageData/SelectAttributesModal/AttributeGroup/AttributeGroup.module.css
