+++ src/pages/ManageData/SelectAttributesModal/SelectAttributesModal.jsx
