--- src/pages/QueryData/QueryResults/QueryResults.module.css
@@ -1,3 +1,0 @@
-.query-results-container{
-  max-width: 95%;
-}
