+++ src/utils/ProtectedRoute.jsx
@@ -1,6 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { Navigate } from 'react-router-dom';
 import { PropTypes, instanceOf } from 'prop-types';
+import LoadingScreen from '../common/LoadingScreen/LoadingScreen';
 import { GSPBackend } from './utils';
 import { withCookies, cookieKeys, Cookies, clearCookies } from './cookie_utils';
 import { refreshToken } from './auth_utils';
@@ -39,7 +40,7 @@ const ProtectedRoute = ({ Component, redirectPath, roles, cookies }) => {
   }, []);
 
   if (isLoading) {
-    return <h1>LOADING...</h1>;
+    return <LoadingScreen />;
   }
 
   if (isAuthenticated) {
